Change authenticator to User.new to work around login/sign up/AuthProvider combinations
