+++ README.md
@@ -39,11 +39,16 @@ This is a [Discourse](http://www.discourse.org/) plugin for authenticating with 
 
 
 ## Local development
-You can develop with Vagrant.  As you develop, copy this repository to the `plugins` folder and restart Rails to see changes.  This also clears the cache of compiled ERB templates (see below).
+You can develop with Vagrant ([see Discourse docs](https://github.com/discourse/discourse/blob/master/docs/VAGRANT.md)).  As you develop, clear the ERB cache, copy this repository to the `plugins` folder and restart Rails to see changes.
 
 Example:
 ```
-rm -rf tmp/cache && rm -rf ./plugins/discourse-omniauth-lti/ && rsync -av --delete /Users/krob/github/mit-teaching-systems-lab/discourse-omniauth-lti ./plugins/ --exclude .git &&  vagrant ssh -c 'cd /vagrant && bundle exec rails s -b 0.0.0.0'
+rm -rf tmp/cache && \
+rm -rf ./plugins/discourse-omniauth-lti/ && \
+rsync -av --exclude .git \
+  ~/github/mit-teaching-systems-lab/discourse-omniauth-lti \
+  ./plugins/ && \
+vagrant ssh -c 'cd /vagrant && bundle exec rails s -b 0.0.0.0'
 ```
 
-The `plugin-third-party.js.erb` file is what ultimately injects the JavaScript needed to show the new login button in the UI.  In development mode, this file will be cached and won't updated if you are rebuilding the plugin on each change.  You can touch it manually, or see https://meta.discourse.org/t/tmp-cache-needs-to-be-manually-cleared-when-developing-plugins/17109/7 or https://github.com/sstephenson/sprockets/issues/563 for other alternatives.
+The `plugin-third-party.js.erb` file is what ultimately injects the JavaScript needed to show the new login button in the UI.  In development mode, this file will be cached and won't updated if you are rebuilding the plugin on each change.  You can touch it manually or just clear the ERB cache on each change (like above).  See https://meta.discourse.org/t/tmp-cache-needs-to-be-manually-cleared-when-developing-plugins/17109 or https://github.com/sstephenson/sprockets/issues/563 for other alternatives.
